+++ CHANGELOG.md
@@ -1,1 +1,2 @@
 # Version 1.0.0
+- Moved Event class and Event functions from dev.pandasystems.pandalib.utils to dev.pandasystems.pandalib.events